requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
chromadb
groq
pandas
fpdf
python-dotenv
